# Conflicts:
#	cs544-202403-project/src/main/java/edu/miu/cs/cs544/domain/Event.java
#	cs544-202403-project/src/main/java/edu/miu/cs/cs544/domain/Schedule.java
#	cs544-202403-project/src/main/resources/application.yml

+++ cs544-202403-project/src/main/java/edu/miu/cs/cs544/domain/Location.java
@@ -12,7 +12,7 @@ public class Location implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long locationId;
+    private Integer locationId;
     private String name;
     private String description;
